requirements.txt
faiss_cpu==1.10.0
huggingface_hub==0.29.3
langdetect==1.0.9
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
Requests==2.32.3
sentence_transformers==4.1.0
streamlit==1.44.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
